fixed edit user page breaking, fixed page breaking when we ade duplicate skill, added prevention of deleting skills linked to user(s)

+++ app.py
@@ -47,13 +47,21 @@ class User(db.Model):
         return f'<User {self.firstname}>'
 
     def save(self):
-        db.session.add(self)
-        db.session.commit()
-        return self.id
+        try:
+            db.session.add(self)
+            db.session.commit()
+            return self.id
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
     def remove(self):
-        db.session.delete(self)
-        db.session.commit()
+        try:
+            db.session.delete(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
 class Skill(db.Model):
     __tablename__ = 'Skill'
@@ -72,13 +80,22 @@ class Skill(db.Model):
         return f'<Skill {self.name}>'
 
     def save(self):
-        db.session.add(self)
-        db.session.commit()
+        try:
+            db.session.add(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
     def remove(self):
-        # TODO prevent delete if skill is lined to any user
-        db.session.delete(self)
-        db.session.commit()
+        try:
+            if db.session.query(UserSkill.skill_id).filter_by(id=self.id).first() is None:
+                # delete if skill is not linked to any user
+                db.session.delete(self)
+                db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
 class UserEducation(db.Model):
     __tablename__ = 'UserEducation'
@@ -99,12 +116,20 @@ class UserEducation(db.Model):
         return f'<Institution {self.institution}>'
 
     def save(self):
-        db.session.add(self)
-        db.session.commit()
+        try:
+            db.session.add(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
     def remove(self):
-        db.session.delete(self)
-        db.session.commit()
+        try:
+            db.session.delete(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
 class UserExperience(db.Model):
     __tablename__ = 'UserExperience'
@@ -124,12 +149,20 @@ class UserExperience(db.Model):
         return f'<Position {self.position}>'
 
     def save(self):
-        db.session.add(self)
-        db.session.commit()
+        try:
+            db.session.add(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
     def remove(self):
-        db.session.delete(self)
-        db.session.commit()
+        try:
+            db.session.delete(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
 class UserSkill(db.Model):
     __tablename__ = 'UserSkill'
@@ -142,16 +175,28 @@ class UserSkill(db.Model):
         super(UserSkill, self).__init__(**kwargs)
     
     def save(self):
-        db.session.add(self)
-        db.session.commit()
+        try:
+            db.session.add(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
 
     def remove(self):
-        db.session.delete(self)
-        db.session.commit()
+        try:
+            db.session.delete(self)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
     
-    def removeAll(self, id):
-        # sql = delete(UserSkill).where(UserSkill.user_id.in_([ids]))
-        sql = delete(UserSkill).where(UserSkill.user_id == id)
+    def removeAll(id):
+        try:
+            # sql = delete(UserSkill).where(UserSkill.user_id.in_([ids]))
+            sql = delete(UserSkill).where(UserSkill.user_id == id)
 
-        db.session.execute(sql)
-        db.session.commit()
+            db.session.execute(sql)
+            db.session.commit()
+        except Exception as e:
+            # logger for errors
+            db.session.close()
